Move ext to lib
