+++ app/src/main/java/com/myth/mama/MainActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.myth.arch.mvvm2.MythView
-import com.myth.mama.base.toast
+import com.myth.arch.mvvm3.toast
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity(), MythView {
