+++ myth-arch-mvvm-lib/src/main/java/com/myth/arch/mvvm3/MythViewModelExt.kt
@@ -1,4 +1,4 @@
-package com.myth.mama.base
+package com.myth.arch.mvvm3
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,16 +8,14 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
-import com.myth.arch.mvvm3.MythViewModel
-import com.myth.mama.performance.TimePerformance
 
 /**
  * Use Block
  */
-internal fun BaseViewModel.useFragment(callback: (Fragment) -> Unit) {
+fun MythViewModel.useFragment(callback: (Fragment) -> Unit) {
     val name = "useFragment"
 
-    val toastData = provider.getViewModelExtData(name) ?: MutableLiveData<(Fragment) -> Unit>()
+    val toastData = getProvider().getViewModelExtData(name) ?: MutableLiveData<(Fragment) -> Unit>()
 
     toastData.value = callback
 
@@ -25,18 +23,18 @@ internal fun BaseViewModel.useFragment(callback: (Fragment) -> Unit) {
         return
     }
 
-    provider.addViewModelExt(name, toastData) { view, data ->
+    getProvider().addViewModelExt(name, toastData) { view, data ->
         data.observe(view.getLifeCycleOwner(), Observer {
             callback(view.getFragment2())
         })
     }
 }
 
-internal fun BaseViewModel.useActivity(callback: (AppCompatActivity) -> Unit) {
+fun MythViewModel.useActivity(callback: (AppCompatActivity) -> Unit) {
     val name = "useActivity"
 
     val toastData =
-        provider.getViewModelExtData(name) ?: MutableLiveData<(AppCompatActivity) -> Unit>()
+        getProvider().getViewModelExtData(name) ?: MutableLiveData<(AppCompatActivity) -> Unit>()
 
     toastData.value = callback
 
@@ -44,7 +42,7 @@ internal fun BaseViewModel.useActivity(callback: (AppCompatActivity) -> Unit) {
         return
     }
 
-    provider.addViewModelExt(name, toastData) { view, data ->
+    getProvider().addViewModelExt(name, toastData) { view, data ->
         data.observe(view.getLifeCycleOwner(), Observer {
             view.getActivity2()?.let {
                 callback(it as AppCompatActivity)
@@ -56,13 +54,10 @@ internal fun BaseViewModel.useActivity(callback: (AppCompatActivity) -> Unit) {
 /**
  * Toast
  */
-internal fun BaseViewModel.toast(text: String) {
+fun MythViewModel.toast(text: String) {
     val name = "toast"
 
-    val timeRecorder = TimePerformance.getTimeRecorder(name)
-    timeRecorder.start()
-
-    val toastData = provider.getViewModelExtData(name) ?: MutableLiveData<String>()
+    val toastData = getProvider().getViewModelExtData(name) ?: MutableLiveData<String>()
 
     toastData.postValue(text)
 
@@ -70,24 +65,22 @@ internal fun BaseViewModel.toast(text: String) {
         return
     }
 
-    provider.addViewModelExt(name, toastData) { view, data ->
+    getProvider().addViewModelExt(name, toastData) { view, data ->
         data.observe(view.getLifeCycleOwner(), Observer {
             val context = view.getContext2() ?: return@Observer
             Toast.makeText(context, it, Toast.LENGTH_LONG).show()
         })
     }
-
-    timeRecorder.print()
 }
 
 /**
  * Navigator
  */
-internal fun BaseViewModel.startActivity(cls: Class<out AppCompatActivity>, args: Bundle? = null) {
+fun MythViewModel.startActivity(cls: Class<out AppCompatActivity>, args: Bundle? = null) {
     val name = "navigator"
 
     val liveData =
-        provider.getViewModelExtData(name) ?: MutableLiveData<Class<out AppCompatActivity>>()
+        getProvider().getViewModelExtData(name) ?: MutableLiveData<Class<out AppCompatActivity>>()
 
     liveData.postValue(cls)
 
@@ -95,7 +88,7 @@ internal fun BaseViewModel.startActivity(cls: Class<out AppCompatActivity>, args
         return
     }
 
-    provider.addViewModelExt(name, liveData) { view, data ->
+    getProvider().addViewModelExt(name, liveData) { view, data ->
         data.observe(view.getLifeCycleOwner(), Observer {
             val context = view.getContext2() ?: return@Observer
             context.startActivity(Intent(context, it).apply {
@@ -107,19 +100,15 @@ internal fun BaseViewModel.startActivity(cls: Class<out AppCompatActivity>, args
     }
 }
 
-internal fun BaseViewModel.startActivityForResult(
+fun MythViewModel.startActivityForResult(
     cls: Class<out AppCompatActivity>,
     args: Bundle? = null,
     requestCode: Int
 ) {
     val name = "navigatorForResult"
 
-
-    val timeRecorder = TimePerformance.getTimeRecorder(name)
-    timeRecorder.start()
-
     val liveData =
-        provider.getViewModelExtData(name) ?: MutableLiveData<Class<out AppCompatActivity>>()
+        getProvider().getViewModelExtData(name) ?: MutableLiveData<Class<out AppCompatActivity>>()
 
     liveData.postValue(cls)
 
@@ -127,7 +116,7 @@ internal fun BaseViewModel.startActivityForResult(
         return
     }
 
-    provider.addViewModelExt(name, liveData) { view, data ->
+    getProvider().addViewModelExt(name, liveData) { view, data ->
         data.observe(view.getLifeCycleOwner(), Observer {
             when (view) {
                 is FragmentActivity -> {
@@ -150,11 +139,9 @@ internal fun BaseViewModel.startActivityForResult(
             }
         })
     }
-
-    timeRecorder.print()
 }
 
-internal fun MythViewModel.finish() {
+fun MythViewModel.finish() {
     val name = "finish"
 
     val liveData = getProvider().getViewModelExtData(name) ?: MutableLiveData<Boolean>()
@@ -181,7 +168,7 @@ internal fun MythViewModel.finish() {
 }
 
 
-internal fun MythViewModel.popBackStack() {
+fun MythViewModel.popBackStack() {
     val name = "popBackStack"
 
     val liveData = getProvider().getViewModelExtData(name) ?: MutableLiveData<Boolean>()
